Move validation into separate class and add invocation to span receipt step
